+++ packages/rnative/src/components/elements/input.tsx
@@ -1,7 +1,10 @@
 import React from 'react';
 import {InputProps, Input as UIInput} from '../ui/input';
 
-interface Props extends InputProps {}
+interface Props extends InputProps {
+  label?: string;
+  name?: string;
+}
 
 const Input = ({...props}: Props) => {
   return <UIInput {...props} />;